requirements.txt
calmsize
h5py
cvxpy
ujson
torchtext
numpy
opacus
torchvision
loguru
scikit-learn
scipy
pillow